solarinsight-api/pyproject.toml
[project]
name = "solarinsight-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.136.0",
    "httpx>=0.28.1",
    "python-dotenv>=1.2.2",
    "uvicorn>=0.34.0",
]
